new file:   FCM_INTEGRATION_GUIDE.md 	new file:   Policy_Renewal_Workflow_Manual.md 	modified:   android/app/build.gradle 	modified:   android/app/build.gradle.kts 	new file:   android/app/google-services.json 	modified:   android/app/src/main/res/mipmap-hdpi/ic_launcher.png 	modified:   android/app/src/main/res/mipmap-mdpi/ic_launcher.png 	modified:   android/app/src/main/res/mipmap-xhdpi/ic_launcher.png 	modified:   android/app/src/main/res/mipmap-xxhdpi/ic_launcher.png 	modified:   android/app/src/main/res/mipmap-xxxhdpi/ic_launcher.png 	modified:   android/build.gradle 	modified:   android/build.gradle.kts 	new file:   assets/images/icon.png 	modified:   ios/Runner.xcodeproj/project.pbxproj 	modified:   ios/Runner/Assets.xcassets/AppIcon.appiconset/Icon-App-1024x1024@1x.png 	modified:   ios/Runner/Assets.xcassets/AppIcon.appiconset/Icon-App-20x20@1x.png 	modified:   ios/Runner/Assets.xcassets/AppIcon.appiconset/Icon-App-20x20@2x.png 	modified:   ios/Runner/Assets.xcassets/AppIcon.appiconset/Icon-App-20x20@3x.png 	modified:   ios/Runner/Assets.xcassets/AppIcon.appiconset/Icon-App-29x29@1x.png 	modified:   ios/Runner/Assets.xcassets/AppIcon.appiconset/Icon-App-29x29@2x.png 	modified:   ios/Runner/Assets.xcassets/AppIcon.appiconset/Icon-App-29x29@3x.png 	modified:   ios/Runner/Assets.xcassets/AppIcon.appiconset/Icon-App-40x40@1x.png 	modified:   ios/Runner/Assets.xcassets/AppIcon.appiconset/Icon-App-40x40@2x.png 	modified:   ios/Runner/Assets.xcassets/AppIcon.appiconset/Icon-App-40x40@3x.png 	new file:   ios/Runner/Assets.xcassets/AppIcon.appiconset/Icon-App-50x50@1x.png 	new file:   ios/Runner/Assets.xcassets/AppIcon.appiconset/Icon-App-50x50@2x.png 	new file:   ios/Runner/Assets.xcassets/AppIcon.appiconset/Icon-App-57x57@1x.png 	new file:   ios/Runner/Assets.xcassets/AppIcon.appiconset/Icon-App-57x57@2x.png 	modified:   ios/Runner/Assets.xcassets/AppIcon.appiconset/Icon-App-60x60@2x.png 	modified:   ios/Runner/Assets.xcassets/AppIcon.appiconset/Icon-App-60x60@3x.png 	new file:   ios/Runner/Assets.xcassets/AppIcon.appiconset/Icon-App-72x72@1x.png 	new file:   ios/Runner/Assets.xcassets/AppIcon.appiconset/Icon-App-72x72@2x.png 	modified:   ios/Runner/Assets.xcassets/AppIcon.appiconset/Icon-App-76x76@1x.png 	modified:   ios/Runner/Assets.xcassets/AppIcon.appiconset/Icon-App-76x76@2x.png 	modified:   ios/Runner/Assets.xcassets/AppIcon.appiconset/Icon-App-83.5x83.5@2x.png 	modified:   lib/core/Transactions/controller/customer_transactions_controller.dart 	modified:   lib/core/Transactions/views/customer_transactions_list_screen.dart 	new file:   lib/core/auth/controller/forgot_password_controller.dart 	new file:   lib/core/auth/data/auth_service.dart 	new file:   lib/core/auth/view/forgot_password_screen.dart 	new file:   lib/core/auth/view/new_password_screen.dart 	new file:   lib/core/auth/view/otp_reset_screen.dart 	modified:   lib/core/claim/controller/claim_controller.dart 	modified:   lib/core/claim/controller/lodge_claim_controller.dart 	modified:   lib/core/claim/views/claim_details_screen.dart 	modified:   lib/core/claim/views/lodge_claims_screen.dart 	modified:   lib/core/claim/widget/main_tab.dart 	new file:   lib/core/landing/controller/landing_controller.dart 	modified:   lib/core/landing/views/landing_screen.dart 	modified:   lib/core/landing/widget/contact_details.dart 	modified:   lib/core/landing/widget/landing_container.dart 	modified:   lib/core/login/view/login_screen.dart 	modified:   lib/core/policy/controller/endorsement_controller.dart 	modified:   lib/core/policy/controller/renew_policy_controller.dart 	modified:   lib/core/policy/views/endorsement_summary_screen.dart 	modified:   lib/core/policy/views/renew_policy_confirmation_screen.dart 	modified:   lib/core/profile/controller/profile_controller.dart 	modified:   lib/core/profile/views/info_screen.dart 	modified:   lib/core/quote/controller/quote_payment_controller.dart 	modified:   lib/core/quote/controller/quote_summary_controller.dart 	modified:   lib/core/quote/views/quote_summary_screen.dart 	modified:   lib/core/quote/widgets/quote_items_list.dart 	new file:   lib/core/services/firebase_messaging_service.dart 	modified:   lib/core/splash/splash_screen.dart 	new file:   lib/firebase_options.dart 	modified:   lib/internalization/app_strings.dart 	modified:   lib/internalization/english_strings.dart 	modified:   lib/main.dart 	modified:   lib/navigation/routes.dart 	modified:   lib/network/api/api_provider.dart 	modified:   lib/network/api/endpoints.dart 	modified:   lib/network/models/request/create_receipt_request.dart 	modified:   lib/network/repository/pibro_repository.dart 	modified:   lib/shared/auth_bg.dart 	new file:   lib/shared/contest_modal.dart 	new file:   lib/shared/widget/app_text_field.dart 	new file:   lib/shared/widget/primary_cta.dart 	new file:   lib/shared/widget/success_dialog.dart 	modified:   lib/utils/app_utils.dart 	modified:   macos/Flutter/GeneratedPluginRegistrant.swift 	modified:   pubspec.lock 	modified:   pubspec.yaml 	modified:   windows/flutter/generated_plugin_registrant.cc 	modified:   windows/flutter/generated_plugins.cmake

+++ lib/network/models/request/create_receipt_request.dart
@@ -14,7 +14,7 @@ class CreateReceiptRequest {
   late String? receiptID;
   late String? transactionDate;
   late String? systemDate;
-  late int? amount;
+  late double? amount;
   late String? documentNumber;
   late String? documentDate;
   late String? channel;
